Lab01-CNN-GTSRB/requirements.txt
python==3.7.16  
matplotlib.pyplot==3.5.3  
numpy==1.21.6  
pandas==1.3.5  
scikit-learn==1.0.2  
torch==1.13.1  